ISSUE #276 * Fix bug related with refactoring.
